# swagger.yaml
openapi: 3.0.0
info:
  title: IMF Gadget API
  version: 1.0.0
  description: |
    Secure API for managing classified IMF gadgets.
    ## Authentication
    - Register agent first at `/register`
    - Get JWT token from `/login`
    - Use `Bearer <token>` in Authorization header for protected endpoints

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://imf-gadget-api-0akv.onrender.com
    description: Production server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Gadget:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: The Nightingale
        status:
          type: string
          enum: [Available, Deployed, Destroyed, Decommissioned]
          example: Available
        missionSuccessProbability:
          type: integer
          minimum: 0
          maximum: 100
          example: 75
        createdAt:
          type: string
          format: date-time
          example: 2024-02-20T12:34:56Z
        decommissionedAt:
          type: string
          format: date-time
          nullable: true
          example: null

tags:
  - name: Authentication
    description: Agent registration and authentication
  - name: Gadgets
    description: Mission-critical gadget operations

paths:
  /register:
    post:
      tags: [Authentication]
      summary: Register new agent
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  example: ethan_hunt
                password:
                  type: string
                  minLength: 6
                  example: impossible
      responses:
        201:
          description: Agent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  username:
                    type: string
        400:
          description: Invalid registration data

  /login:
    post:
      tags: [Authentication]
      summary: Authenticate agent
      operationId: loginAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: ethan_hunt
                password:
                  type: string
                  example: impossible
      responses:
        200:
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjEyMzQ1IiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        401:
          description: Invalid credentials

  /gadgets:
    get:
      tags: [Gadgets]
      summary: List all gadgets
      operationId: getGadgets
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [Available, Deployed, Destroyed, Decommissioned]
          description: Filter gadgets by status
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gadget'
        500:
          description: Server error

    post:
      tags: [Gadgets]
      summary: Create new gadget
      operationId: createGadget
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 3
                  example: Grappling Hook
      responses:
        201:
          description: Gadget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gadget'
        400:
          description: Invalid input
        401:
          description: Unauthorized

  /gadgets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000

    patch:
      tags: [Gadgets]
      summary: Update gadget
      operationId: updateGadget
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 3
                status:
                  type: string
                  enum: [Available, Deployed, Destroyed, Decommissioned]
      responses:
        200:
          description: Gadget updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gadget'
        400:
          description: Invalid input
        401:
          description: Unauthorized
        404:
          description: Gadget not found

    delete:
      tags: [Gadgets]
      summary: Decommission gadget
      operationId: decommissionGadget
      security:
        - BearerAuth: []
      responses:
        200:
          description: Gadget decommissioned
        401:
          description: Unauthorized
        404:
          description: Gadget not found

  /gadgets/{id}/self-destruct:
    post:
      tags: [Gadgets]
      summary: Initiate self-destruct
      operationId: selfDestructGadget
      security:
        - BearerAuth: []
      responses:
        200:
          description: Destruct sequence initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  confirmationCode:
                    type: string
                    example: X9FG3H
        401:
          description: Unauthorized
        404:
          description: Gadget not found